fix status bar height calculation endianess issue (#1484)
* fix status bar height calculation endianess issue

Fixes #1482

* fix endianess issue in V_DrawPatchInt()

+++ src/v_video.c
@@ -524,7 +524,7 @@ static void V_DrawPatchInt(int x, int y, patch_t *patch, boolean flipped)
             }
     #endif
 
-            column = (column_t *)((byte *)patch + patch->columnofs[texturecolumn]);
+            column = (column_t *)((byte *)patch + LONG(patch->columnofs[texturecolumn]));
             V_DrawMaskedColumn(&patchcol, ytop, column);
         }
     }
